feat: listen to file dnd event on window instead of just popover
